+++ src/main/java/com/crudshop/demo/dto/OrderDto.java
@@ -5,6 +5,7 @@ import com.crudshop.demo.entity.OrderStatus;
 import lombok.Builder;
 import lombok.Value;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Value
@@ -15,7 +16,7 @@ public class OrderDto {
 
     CustomerEntity customer;
 
-    Double totalPrice;
+    BigDecimal totalPrice;
 
     OrderStatus status;
