+++ src/main/java/com/crudshop/demo/interaction/ExchangeRateClientImplStub.java
@@ -7,6 +7,8 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
+
 @Component
 @Primary
 @Slf4j
@@ -15,7 +17,7 @@ public class ExchangeRateClientImplStub implements ExchangeRateClient {
     @Override
     public ExchangeRate getExchangeRate() {
         ExchangeRate exchangeRate = new ExchangeRate();
-        exchangeRate.setExchangeRateRUB(20.0);
+        exchangeRate.setExchangeRateRUB(BigDecimal.valueOf(20.0));
         log.info("Значение получено из стаба");
 
         return exchangeRate;
